requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
pymupdf
python-dotenv
groq